pyproject.toml
[tool.poetry]
name = "Depth-Aware-Object-Detection"
version = "0.1.0"
description = "Depth-Aware Object Detection for Autonomous Urban Navigation"
authors = ["Maryam Mohebbi <maryam.mohebbi.2@gmail.com>"]
readme = "README.md"
packages = [{ include = "da_od" }]

[tool.poetry.dependencies]
python = "3.10.10"
numpy = "^1.26.2"
pandas = "^2.1.3"
matplotlib = "^3.8.2"
scipy = "^1.11.3"
scikit-learn = "^1.3.2"
jupyter = "^1.0.0"
seaborn = "^0.13.0"
plotly = "^5.18.0"
torch = "^2.1.1"
pytorch-lightning = "^2.1.2"
pytorch-forecasting = "^1.0.0"
tensorboardX = "^2.6.2.2"
tensorboard = "^2.15.1"
ipykernel = "^6.26.0"
ipython = "^8.17.2"
jupyterlab = "^4.0.8"
ipywidgets = "^8.1.1"
SQLAlchemy = "^2.0.22"
tensorflow = "^2.15.0"
keras = ">=2.15.0,<2.16"
opencv-python = "^4.8.1.78"
torchvision = "^0.16.1"
pillow = "^10.1.0"
timm = "^0.9.12"

[tool.poetry.group.dev.dependencies]
mypy = "^1.7.0"
ruff = "^0.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.10"
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.black]
line-length = 110

[tool.ruff]
target-version = "py39"
line-length = 110
fix = true
select = ["ALL"]
ignore = ["D100", "D104", "ISC003", "TD003", "EXE002", "PLR2004", "PLR0911", "PLR0913"]

[tool.ruff.pydocstyle]
convention = "google"